+++ src/neo/Network/P2P/Payloads/Conditions/OrCondition.cs
@@ -8,47 +8,48 @@
 // Redistribution and use in source and binary forms with or without
 // modifications are permitted.
 
-using Neo.Cryptography.ECC;
 using Neo.IO;
 using Neo.IO.Json;
 using Neo.SmartContract;
-using Neo.SmartContract.Native;
+using System;
 using System.IO;
 using System.Linq;
 
-namespace Neo.Network.P2P.Payloads
+namespace Neo.Network.P2P.Payloads.Conditions
 {
-    public class GroupCondition : WitnessCondition
+    /// <summary>
+    /// Represents the condition that any of the conditions meets.
+    /// </summary>
+    public class OrCondition : WitnessCondition
     {
         /// <summary>
-        /// The group to be checked.
+        /// The expressions of the condition.
         /// </summary>
-        public ECPoint Group;
+        public WitnessCondition[] Expressions;
 
-        public override int Size => base.Size + Group.Size;
-        public override WitnessConditionType Type => WitnessConditionType.Group;
+        public override int Size => base.Size + Expressions.GetVarSize();
+        public override WitnessConditionType Type => WitnessConditionType.Or;
 
         protected override void DeserializeWithoutType(BinaryReader reader)
         {
-            Group = reader.ReadSerializable<ECPoint>();
+            Expressions = DeserializeConditions(reader);
+            if (Expressions.Length == 0) throw new FormatException();
         }
 
         public override bool Match(ApplicationEngine engine)
         {
-            engine.ValidateCallFlags(CallFlags.ReadStates);
-            ContractState contract = NativeContract.ContractManagement.GetContract(engine.Snapshot, engine.CurrentScriptHash);
-            return contract is not null && contract.Manifest.Groups.Any(p => p.PubKey == Group);
+            return Expressions.Any(p => p.Match(engine));
         }
 
         protected override void SerializeWithoutType(BinaryWriter writer)
         {
-            writer.Write(Group);
+            writer.Write(Expressions);
         }
 
         public override JObject ToJson()
         {
             JObject json = base.ToJson();
-            json["group"] = Group.ToString();
+            json["expressions"] = Expressions.Select(p => p.ToJson()).ToArray();
             return json;
         }
     }
