+++ src/neo/Network/P2P/Payloads/Signer.cs
@@ -11,6 +11,7 @@
 using Neo.Cryptography.ECC;
 using Neo.IO;
 using Neo.IO.Json;
+using Neo.Network.P2P.Payloads.Conditions;
 using System;
 using System.Collections.Generic;
 using System.IO;
