+++ src/neo/Network/P2P/Payloads/WitnessRule.cs
@@ -10,6 +10,7 @@
 
 using Neo.IO;
 using Neo.IO.Json;
+using Neo.Network.P2P.Payloads.Conditions;
 using System;
 using System.IO;
